backend/requirements.txt
flask
flask-socketio
eventlet
flask_sqlalchemy
flask-migrate
requests
APScheduler>=3.10.0
pytz
flask-cors 
python-dotenv 